src/requirements.txt
discord.py
pytesseract
opencv-python
RestrictedPython